requirements.txt
PyYAML==6.0
pyserial==3.5
